correction methode division double

+++ Calcul.java
@@ -17,7 +17,7 @@ public class Calcul {
 
     public static double division(int n1, int n2){
         if (n2 != 0)
-            return n1 / n2;
+            return (double)n1 / (double)n2;
         else
             return 0;
     }
